% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{connections}
\alias{connections}
\title{API "connections" endpoint}
\usage{
connections(..., .protocol = "https")
}
\arguments{
\item{...}{\code{from} and \code{to} are required}

\item{.protocol}{the connection protocol; either http or https (default)}
}
\value{
the connections as a dataframe
}
\description{
For detailed api documentation see 
    \url{https://transport.opendata.ch/docs.html#connections}
}
\examples{
connections(from = "Bern", to = "Burgdorf")
connections(from = "Basel", to = "Bern")

}
