% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwissPublicTransport.R
\docType{package}
\name{SwissPublicTransport-package}
\alias{SwissPublicTransport}
\alias{SwissPublicTransport-package}
\title{swiss public transport api R client}
\description{
This is a client for the Swiss public transport API.
}
\author{
\strong{Maintainer}: Reto Wyss \email{reto.wyss.ch@gmail.com}

}
