% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapi.R
\name{tapi}
\alias{tapi}
\title{Swiss public transport API}
\usage{
tapi(endpoint, ..., .protocol = "https", .url = "transport.opendata.ch",
  .version = "v1")
}
\arguments{
\item{endpoint}{path to api endpoint - there are three endpoints available:
locations, connections, and stationboard.}

\item{...}{named parameters will be translated into a request query}

\item{.protocol}{protocol to be used}

\item{.url}{base url}

\item{.version}{api version}
}
\value{
list of dataframes
}
\description{
Swiss public transport API
}
\examples{

tapi("locations", query = "Bern")
tapi("connections", from = "Bern", to = "Burgdorf")
tapi("stationboard", station = "Burgdorf")

}
