% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport_api.R
\name{transport_api}
\alias{transport_api}
\title{Swiss public transport API}
\usage{
transport_api(.endpoint, ..., .protocol = "https", .version = "v1")
}
\arguments{
\item{.endpoint}{path to api endpoint - there are three endpoints available:
locations, connections, and stationboard.}

\item{...}{named parameters will be translated into a request query.
For the full documentation see \url{https://transport.opendata.ch/docs.html}}

\item{.protocol}{protocol to be used}

\item{.version}{api version}
}
\value{
list of dataframes
}
\description{
Swiss public transport API
}
